Add files via upload
This program is designed to take in an .txt file which contains a list of .txt file names along with an shingle increment through the command prompt. It will take the increment and use it to compare each of the text files to the others then print out a ratio and a percentage to represent their similarity.